cogs: command_create.py: move buttons to another file
